Change all instances of `Group` to `Sponsor`
Summary:
The idea is to unify the concepts of "groups" and "independent sponsors".
Functionally on the backend, they already both operate under the "group" model.
This simply changes the naming to make it more clear and allow the single
Sponsor concept to be exposed in the UI as well. Consistency!

## Outstanding Questions

* The key and constraint names in the DB still have `group` in their names

Closes T162

Test Plan:
* Run `make test`
* Make sure sponsors show in document lists
* Make sure you can create new documents as a sponsor
* Make sure you can change a document's sponsor
* Test other various sponsor related functionality?
* **Bonus points:** Try features we haven't reimplemented yet in the console!

Reviewers: doshitan

Maniphest Tasks: T162

Differential Revision: https://phabricator.opengovfoundation.org/D62