Notify sponsors when a comment is created on their documents
Summary:
Sponsor models themselves could implement the Notifiable trait (or specifically
the `routeNotificationFor` method). But I'm not sure how good the support of
this is for delivery methods that aren't email and this sends a single email to
all the addresses (instead of individual emails to each) so how this interacts
with a user setting about the notification or filtering a users own actions
would be complicated. So for now, we'll just build the list of user entities by
hand.

Test Plan:
- Have at least two members in a sponsor
- Comment as a user on a document of that sponsor
- Ensure the sponors recieve an email

Reviewers: sethetter

Subscribers: