Document search
Summary:
Add fulltext indexes to document titles and content and utilize them for
searching. The search works in boolean mode so basic stuff (quotes for exact
match, not operator, etc.) works, should be enough for our needs for now.

MySQL by default ORs search terms together, so things like 'red car' will search
for anything that possess the term 'red' or 'car'. Generally folks want their
terms ANDed together so 'red car' means search for things that have both 'red'
and 'car' in them. You have to do this at the application level for InnoDB
tables. That's what `SearchQueryCompiler` is for, it will AND all the terms
together unless the user has specified otherwise (they can say OR with a '|').
It also wildcards the terms which is another thing that is usually desirable
when searching. If the user doesn't want a term wildcarded, they can quote it.
It does not recurse into subexpressions. We could do that if we wanted, probably
unnecessary though.

Note we never remove indexed content since we never actually delete documents
and their content from the database, just soft delete stuff. So it will always
just kinda grow in size, could become something to worry about in the far
future. Might still be worth it to run an optimize or otherwise reindex the
stuff every so often, but I'm not thinking it will be an issue at the moment.

Also:
- Move document search field from home page to navbar
- Fix pagination template

Test Plan:
- Run the migrations
- Search for some documents

Reviewers: sethetter

Maniphest Tasks: T230

Differential Revision: https://phabricator.opengovfoundation.org/D127